fixed CreateGame
